Extract decorateNameMatchedToSearchTerm to utils
