+++ lib/ui/src/modules/ui/components/left_panel/stories_tree/index.js
@@ -136,7 +136,7 @@ class Stories extends React.Component {
         name: story.name,
         story: story.name,
         kind: storiesHierarchy.kind,
-        active: selectedStory === story && selectedKind === storiesHierarchy.kind,
+        active: selectedStory === story.name && selectedKind === storiesHierarchy.kind,
         type: treeNodeTypes.STORY,
         highlight: story.highlight,
       }));
