+++ lib/ui/src/modules/ui/components/left_panel/stories_tree/tree_decorators.js
@@ -5,6 +5,7 @@ import PropTypes from 'prop-types';
 import RoutedLink from '../../../containers/routed_link';
 import MenuItem from '../../menu_item';
 import treeNodeTypes from './tree_node_type';
+import { decorateNameMatchedToSearchTerm } from './tree_decorators_utils';
 
 function noop() {}
 
@@ -84,89 +85,29 @@ ContainerDecorator.propTypes = {
   onClick: PropTypes.func.isRequired,
 };
 
-class HeaderDecorator extends React.Component {
-  getParts(name, highlight) {
-    const nameParts = [];
-    let last = 0;
+function HeaderDecorator(props) {
+  const { style, node, ...restProps } = props;
 
-    highlight.forEach(([start, end]) => {
-      if (last < start) {
-        nameParts.push({
-          strong: false,
-          text: name.substring(last, start),
-        });
-      }
+  let newStyle = style;
 
-      nameParts.push({
-        strong: true,
-        text: name.substring(start, end + 1),
-      });
-
-      last = end + 1;
-    });
-
-    if (last < name.length) {
-      nameParts.push({
-        strong: false,
-        text: name.substring(last, name.length),
-      });
-    }
-
-    return nameParts;
-  }
-
-  decorateNameMatchedToSearchTerm(node, style) {
-    const { name, highlight } = node;
-
-    if (!highlight) {
-      return name;
-    }
-
-    const nameParts = this.getParts(name, highlight);
-
-    return nameParts.filter(part => part.text).map((part, index) => {
-      const key = `${part.text}-${index}`;
-
-      if (part.strong) {
-        return (
-          <strong key={key} style={style.highLightText}>
-            {part.text}
-          </strong>
-        );
-      }
-
-      return (
-        <span key={key}>
-          {part.text}
-        </span>
-      );
-    });
-  }
-
-  render() {
-    const { style, node, ...restProps } = this.props;
-
-    let newStyle = style;
-
-    if (node.type === treeNodeTypes.STORY) {
-      newStyle = {
-        ...style,
-        title: {
-          ...style.title,
-          ...style.storyTitle,
-        },
-      };
-    }
-
-    const name = this.decorateNameMatchedToSearchTerm(node, style);
-
-    const newNode = {
-      ...node,
-      name,
+  if (node.type === treeNodeTypes.STORY) {
+    newStyle = {
+      ...style,
+      title: {
+        ...style.title,
+        ...style.storyTitle,
+      },
     };
-
-    return <decorators.Header style={newStyle} node={newNode} {...restProps} />;
   }
+
+  const name = decorateNameMatchedToSearchTerm(node, style);
+
+  const newNode = {
+    ...node,
+    name,
+  };
+
+  return <decorators.Header style={newStyle} node={newNode} {...restProps} />;
 }
 
 HeaderDecorator.propTypes = {
